clean up auth packages
