--- artifacts/checksums.txt
@@ -1,8 +1,0 @@
-748a572cdbb35aeb54fa3fb18665651eaad7ca0332b663da3c85bc3b9ed1bf62  btsg_account_backup.wasm
-3d619b22d5566784dcc64eaccaabbc0f2a0b325b48c9fe64c0fa41e0fe44c3d3  btsg_dao.wasm
-a601a7aabea803bbb9c0454ea7b5d82c933a130d1a456279a17d39a27ace1757  btsg_ed25519.wasm
-9c5e2b454da5eead278356bdc8fcebdb2e86bba3730e96fc025bdd9a693904dd  btsg_eth.wasm
-e21563260c03cfdab1f195878f49bf93722027ad26fcd097cfebbc5c4d279082  btsg_irl.wasm
-fad747ffd9f7250304a5f9b05429dd6b218f166bd329c171a0551638560375c7  btsg_passkey.wasm
-62128bad45d65d53a32f8f9326db62f7aa7a2792894c77f1b83903010ec58314  btsg_wavs.wasm
-3fc113beb246ed1053d3dea27b26e0c0b0be0a40ad8cd565d44a50196bfbb2d0  btsg_zktls.wasm
