--- contracts/smart-accounts/btsg-ed25519/src/state.rs
@@ -1,4 +1,0 @@
-use cw_storage_plus::Item;
-
-/// Stores the contract's configuration
-pub const PUBLIC_KEY: Item<String> = Item::new("pk");
