--- contracts/smart-accounts/btsg-eth/src/contract.rs
@@ -1,96 +1,0 @@
-use btsg_account::traits::default::BtsgAccountTrait;
-
-use cosmwasm_std::entry_point;
-use cosmwasm_std::{Binary, Deps, DepsMut, Env, MessageInfo, Response, StdResult};
-use cw2::set_contract_version;
-
-use crate::BtsgAccountEth;
-use crate::{
-    state::PUBLIC_KEY,
-    ContractError, {ExecuteMsg, InstantiateMsg, QueryMsg, SudoMsg},
-};
-
-// version info for migration info
-const CONTRACT_NAME: &str = "crates.io:btsg-eth";
-const CONTRACT_VERSION: &str = env!("CARGO_PKG_VERSION");
-
-/// Can only be called by governance
-#[cfg_attr(not(feature = "library"), entry_point)]
-pub fn instantiate(
-    deps: DepsMut,
-    _env: Env,
-    info: MessageInfo,
-    msg: InstantiateMsg,
-) -> Result<Response, ContractError> {
-    set_contract_version(deps.storage, CONTRACT_NAME, CONTRACT_VERSION)?;
-
-    PUBLIC_KEY.save(deps.storage, &msg.pubkey)?;
-    cw_ownable::initialize_owner(
-        deps.storage,
-        deps.api,
-        Some(msg.owner.unwrap_or(info.sender).as_str()),
-    )?;
-
-    Ok(Response::new())
-}
-
-#[cfg_attr(not(feature = "library"), entry_point)]
-pub fn query(_deps: Deps, _env: Env, msg: QueryMsg) -> StdResult<Binary> {
-    match msg {}
-}
-
-#[cfg_attr(not(feature = "library"), entry_point)]
-pub fn execute(
-    _deps: DepsMut,
-    _env: Env,
-    _info: MessageInfo,
-    msg: ExecuteMsg,
-) -> Result<Response, ContractError> {
-    match msg {}
-}
-
-#[cfg_attr(not(feature = "library"), entry_point)]
-pub fn sudo(deps: DepsMut, env: Env, msg: SudoMsg) -> Result<Response, ContractError> {
-    BtsgAccountEth::process_sudo_auth(deps, env, &msg)
-}
-
-pub fn execute_update_owner(
-    deps: DepsMut,
-    info: MessageInfo,
-    env: Env,
-    action: cw_ownable::Action,
-) -> Result<Response, ContractError> {
-    let ownership = cw_ownable::update_ownership(deps, &env.block, &info.sender, action)?;
-    Ok(Response::default().add_attributes(ownership.into_attributes()))
-}
-
-#[cfg(test)]
-mod test {
-    use cosmwasm_std::testing::mock_dependencies;
-    use saa::EthPersonalSign;
-    use saa_common::Verifiable;
-
-    #[test]
-    fn eth_personal_verifiable() {
-        let deps = mock_dependencies();
-
-        let message = r#"{"chain_id":"elgafar-1","contract_address":"stars1gjgfp9wps9c0r3uqhr0xxfgu02rnzcy6gngvwpm7a78j7ykfqquqr2fuj4","messages":["Create TBA account"],"nonce":"0"}"#;
-        let address = "0xac03048da6065e584d52007e22c69174cdf2b91a";
-        let base = "eyJjaGFpbl9pZCI6ImVsZ2FmYXItMSIsImNvbnRyYWN0X2FkZHJlc3MiOiJzdGFyczFnamdmcDl3cHM5YzByM3VxaHIweHhmZ3UwMnJuemN5NmduZ3Z3cG03YTc4ajd5a2ZxcXVxcjJmdWo0IiwibWVzc2FnZXMiOlsiQ3JlYXRlIFRCQSBhY2NvdW50Il0sIm5vbmNlIjoiMCJ9";
-        let message = cosmwasm_std::Binary::new(message.as_bytes().to_vec());
-        assert!(message.to_base64() == base, "not euqal");
-
-        let signature = cosmwasm_std::Binary::from_base64(
-            "a/lQuaTyhcTEeRA2XFTPxoDSIdS3yUUH1VSKOm2zz5EURfheGzzLgXea6QAalswOM2njnUzblqIGiOC0P+j2rhw="
-        ).unwrap();
-
-        let cred = EthPersonalSign {
-            signer: address.to_string(),
-            signature: signature.clone(),
-            message,
-        };
-        let res = cred.verify(deps.as_ref());
-        println!("Res: {:?}", res);
-        assert!(res.is_ok())
-    }
-}
