--- contracts/smart-accounts/btsg-passkey/src/contract.rs
@@ -1,64 +1,0 @@
-use btsg_account::traits::default::BtsgAccountTrait;
-use cosmwasm_std::{Binary, Deps, DepsMut, Env, MessageInfo, Response, StdResult};
-use cw2::set_contract_version;
-
-use crate::{
-    state::PAYLOAD,
-    BtsgAccountPasskey, ContractError, {ExecuteMsg, InstantiateMsg, QueryMsg, SudoMsg},
-};
-
-use cosmwasm_std::entry_point;
-
-// version info for migration info
-const CONTRACT_NAME: &str = "crates.io:btsg-passkey";
-const CONTRACT_VERSION: &str = env!("CARGO_PKG_VERSION");
-
-/// Can only be called by governance
-#[cfg_attr(not(feature = "library"), entry_point)]
-pub fn instantiate(
-    deps: DepsMut,
-    _env: Env,
-    info: MessageInfo,
-    msg: InstantiateMsg,
-) -> Result<Response, ContractError> {
-    set_contract_version(deps.storage, CONTRACT_NAME, CONTRACT_VERSION)?;
-
-    PAYLOAD.save(deps.storage, &msg.payload)?;
-    cw_ownable::initialize_owner(
-        deps.storage,
-        deps.api,
-        Some(msg.owner.unwrap_or(info.sender).as_str()),
-    )?;
-
-    Ok(Response::new())
-}
-
-#[cfg_attr(not(feature = "library"), entry_point)]
-pub fn query(_deps: Deps, _env: Env, msg: QueryMsg) -> StdResult<Binary> {
-    match msg {}
-}
-
-#[cfg_attr(not(feature = "library"), entry_point)]
-pub fn execute(
-    _deps: DepsMut,
-    _env: Env,
-    _info: MessageInfo,
-    msg: ExecuteMsg,
-) -> Result<Response, ContractError> {
-    match msg {}
-}
-
-#[cfg_attr(not(feature = "library"), entry_point)]
-pub fn sudo(deps: DepsMut, env: Env, msg: SudoMsg) -> Result<Response, ContractError> {
-    BtsgAccountPasskey::process_sudo_auth(deps, env, &msg)
-}
-
-pub fn execute_update_owner(
-    deps: DepsMut,
-    info: MessageInfo,
-    env: Env,
-    action: cw_ownable::Action,
-) -> Result<Response, ContractError> {
-    let ownership = cw_ownable::update_ownership(deps, &env.block, &info.sender, action)?;
-    Ok(Response::default().add_attributes(ownership.into_attributes()))
-}
